+++ choicesenum/django/fields.py
@@ -94,6 +94,15 @@ class EnumFieldMixin(object):
                 del kwargs["choices"]
         return name, path, args, kwargs
 
+    def south_field_triple(self):  # pragma: no cover
+        "Returns a suitable description of this field for South (Django 1.6)"
+        from south.modelsinspector import introspector
+        path = "%s.%s" % (self.__class__.__module__, self.__class__.__name__)
+        args, kwargs = introspector(self)
+        if 'default' in kwargs and self.default:
+            kwargs['default'] = repr(self.to_python(self.default).value)
+        return (str(path), args, kwargs)
+
 
 def get_base_classes(cls_type):  # pragma: no cover, already covered by tox matrix
     if hasattr(models, 'SubfieldBase'):
@@ -111,15 +120,3 @@ class EnumCharField(get_base_classes(models.CharField)):
 
 class EnumIntegerField(get_base_classes(models.IntegerField)):
     description = "An integer enum field"
-
-
-try:  # pragma: no cover
-    # if south is installed, be nice and let it know that it has the same
-    # introspection rules as its base class
-    from south.modelsinspector import add_introspection_rules
-    add_introspection_rules(
-        [], ["^choicesenum\.django\.fields\.EnumCharField"])
-    add_introspection_rules(
-        [], ["^choicesenum\.django\.fields\.EnumIntegerField"])
-except ImportError:
-    pass