+++ choicesenum/enums.py
@@ -22,6 +22,9 @@ class ChoicesEnum(Enum):
     def __str__(self):
         return str(self.value)
 
+    def __repr__(self):
+        return "%s(%r).%s" % (self.__class__.__name__, self._value_, self._name_, )
+
     def __eq__(self, other):
         return self.value == getattr(other, 'value', other)